+++ src/main/kotlin/de/ott/ivy/ui/overview/event/IvyLeeEventHandler.kt
@@ -1,11 +1,13 @@
 package de.ott.ivy.ui.overview.event
 
-import de.ott.ivy.data.enums.TaskStatus
+import de.ott.ivy.ui.dialog.TaskDialog
 import de.ott.ivy.ui.overview.IvyLee
 import de.ott.ivy.ui.overview.IvyLee.Companion.getCellByBorderPane
 import de.ott.ivy.ui.overview.IvyLee.Companion.getTaskByBorderPane
-import de.ott.ivy.ui.dialog.TaskDialog
+import de.ott.ivy.ui.overview.event.impl.BrowserHandler
+import de.ott.ivy.ui.overview.event.impl.DeleteButtonHandler
 import de.ott.ivy.ui.overview.impl.TaskCellUpdater
+import javafx.scene.image.ImageView
 import javafx.scene.input.MouseEvent
 import javafx.scene.layout.BorderPane
 import javafx.scene.layout.VBox
@@ -32,46 +34,16 @@ class IvyLeeEventHandler(private val taskList: VBox) {
     }
 
     fun onClick(event: MouseEvent) {
-        if (event.isPrimaryButtonDown)
-            (event.target as BorderPane).apply {
-                val task = IvyLee.tasks.getTaskByBorderPane(this)!!
-                when (task.status) {
-                    TaskStatus.EMPTY -> {
-                        return
-                    }
-                    TaskStatus.UNDONE -> task.status = TaskStatus.IN_WORK
-                    TaskStatus.IN_WORK -> task.status = TaskStatus.DONE
-                    TaskStatus.DONE -> task.status = TaskStatus.UNDONE
-                }
-                TaskCellUpdater.updateTaskCell(task, IvyLee.tasks.getCellByBorderPane(this))
-                if (task.status == TaskStatus.IN_WORK) {
-                    // stop other timers
-                    IvyLee.tasks.values.filter { it.status == TaskStatus.IN_WORK && it != task }.forEach {
-                        it.status = TaskStatus.UNDONE
-                    }
-                    Thread {
-                        val threadCell = IvyLee.tasks.getCellByBorderPane(this)
-                        while (task.status == TaskStatus.IN_WORK) {
-                            task.timeInvestedSeconds++
-                            TaskCellUpdater.updateTaskCell(task, threadCell)
-                            Thread.sleep(1000) // sleep 1s
-                        }
-                    }.start()
-                }
+        when(val element = event.target){
+            is BorderPane -> {
+                if (event.isPrimaryButtonDown)
+                    BrowserHandler.onLeftClick(element)
+                else if(event.isSecondaryButtonDown)
+                    BrowserHandler.onRightClick(element)
             }
-        else {
-            // open custom dialog
-            (event.target as BorderPane).apply {
-                val newTask = TaskDialog.showDialog(IvyLee.tasks.getTaskByBorderPane(this)!!)
-                if (newTask.status == TaskStatus.EMPTY) {
-                    val cell = IvyLee.tasks.getCellByBorderPane(this)
-                    taskList.children.remove(cell.borderPane)
-                    IvyLee.tasks.remove(cell)
-                } else {
-                    IvyLee.tasks[IvyLee.tasks.getCellByBorderPane(this)] = newTask
-                    TaskCellUpdater.updateTaskCell(IvyLee.tasks.getTaskByBorderPane(this)!!, IvyLee.tasks.getCellByBorderPane(this))
-                    mark(event)
-                }
+            is ImageView -> {
+                if(event.isPrimaryButtonDown && element.id == "delete")
+                    DeleteButtonHandler.onLeftClick()
             }
         }
     }
