+++ src/main/kotlin/de/ott/ivy/ui/overview/event/impl/DeleteButtonHandler.kt
@@ -1,0 +1,9 @@
+package de.ott.ivy.ui.overview.event.impl
+
+object DeleteButtonHandler {
+
+    fun onLeftClick(){
+
+    }
+
+}
